@@ -5,7 +5,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.7.0</version>
+        <version>3.3.4</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>com.codeages</groupId>
@@ -14,10 +14,10 @@
     <name>Termius Plus</name>
     <description>Termius Plus</description>
     <properties>
-        <java.version>11</java.version>
+        <java.version>21</java.version>
         <org.mapstruct.version>1.4.2.Final</org.mapstruct.version>
-        <org.projectlombok.version>1.18.22</org.projectlombok.version>
-        <org.apache.maven.compiler-plugin>3.10.1</org.apache.maven.compiler-plugin>
+        <org.projectlombok.version>1.18.34</org.projectlombok.version>
+        <org.apache.maven.compiler-plugin>3.13.0</org.apache.maven.compiler-plugin>
     </properties>
     <dependencies>
         <dependency>
@@ -59,9 +59,9 @@
         </dependency>
 
         <dependency>
-            <groupId>mysql</groupId>
-            <artifactId>mysql-connector-java</artifactId>
-            <scope>runtime</scope>
+            <groupId>com.mysql</groupId>
+            <artifactId>mysql-connector-j</artifactId>
+            <version>9.0.0</version>
         </dependency>
         <dependency>
             <groupId>org.projectlombok</groupId>
@@ -72,7 +72,7 @@
         <dependency>
             <groupId>cn.hutool</groupId>
             <artifactId>hutool-all</artifactId>
-            <version>5.7.22</version>
+            <version>5.8.32</version>
         </dependency>
         <dependency>
             <groupId>org.apache.commons</groupId>
@@ -87,14 +87,17 @@
         <dependency>
             <groupId>com.querydsl</groupId>
             <artifactId>querydsl-apt</artifactId>
-            <version>5.0.0</version>
-            <classifier>jpa</classifier>
+            <version>5.1.0</version>
             <scope>provided</scope>
+            <optional>true</optional>
+            <classifier>jakarta</classifier>
         </dependency>
         <dependency>
+<!--            jakarta-->
             <groupId>com.querydsl</groupId>
             <artifactId>querydsl-jpa</artifactId>
-            <version>5.0.0</version>
+            <version>5.1.0</version>
+            <classifier>jakarta</classifier>
         </dependency>
         <dependency>
             <groupId>org.mapstruct</groupId>
@@ -150,12 +153,12 @@
         <dependency>
             <groupId>org.apache.guacamole</groupId>
             <artifactId>guacamole-common</artifactId>
-            <version>1.5.4</version>
+            <version>1.5.5</version>
         </dependency>
         <dependency>
             <groupId>org.apache.guacamole</groupId>
             <artifactId>guacamole-common-js</artifactId>
-            <version>1.5.4</version>
+            <version>1.5.5</version>
             <type>zip</type>
             <scope>runtime</scope>
         </dependency>
@@ -173,6 +176,16 @@
             <groupId>com.dingtalk.open</groupId>
             <artifactId>dingtalk-stream</artifactId>
             <version>1.3.7</version>
+        </dependency>
+        <dependency>
+            <groupId>javax.xml.bind</groupId>
+            <artifactId>jaxb-api</artifactId>
+            <version>2.3.1</version>
+        </dependency>
+        <dependency>
+            <groupId>org.openjdk.nashorn</groupId>
+            <artifactId>nashorn-core</artifactId>
+            <version>15.4</version>
         </dependency>
     </dependencies>
 
