@@ -19,8 +19,8 @@ import com.codeages.termiusplus.biz.util.TreeUtils;
 import com.codeages.termiusplus.exception.AppException;
 import org.springframework.stereotype.Service;
 
-import javax.transaction.Transactional;
-import javax.validation.Validator;
+import jakarta.transaction.Transactional;
+import jakarta.validation.Validator;
 import java.util.*;
 import java.util.stream.Collectors;
 
