@@ -3,17 +3,15 @@ package com.codeages.termiusplus.biz.objectlog.entity;
 import com.vladmihalcea.hibernate.type.json.JsonStringType;
 import lombok.Data;
 import org.hibernate.annotations.Type;
-import org.hibernate.annotations.TypeDef;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.util.Map;
 
 @Entity
 @EntityListeners(AuditingEntityListener.class)
 @Data
-@TypeDef(name = "json", typeClass = JsonStringType.class)
 public class ObjectLog {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -29,8 +27,7 @@ public class ObjectLog {
 
     private String message;
 
-    @Type(type = "json")
-    private Map<String, Object> context;
+    private String context;
 
     private Long operatorId = 0L;
 
