@@ -17,7 +17,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import javax.validation.Validator;
+import jakarta.validation.Validator;
 
 @Service
 public class CommandServiceImpl implements CommandService {
