@@ -2,7 +2,7 @@ package com.codeages.termiusplus.biz.user.dto;
 
 import lombok.Data;
 
-import javax.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotEmpty;
 import java.util.List;
 
 @Data
