@@ -18,7 +18,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import javax.validation.Validator;
+import jakarta.validation.Validator;
 import java.util.List;
 
 @Service
